Bugfix KNOWAGE-7601
